requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
matplotlib
scikit-learn

